requirements.txt
autopep8==2.0.1
bcrypt==4.0.1
cachelib==0.10.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
pycodestyle==2.10.0
pymongo==4.3.3
python-dotenv==0.21.1
tomli==2.0.1
Werkzeug==2.2.2
